#625 refactor(common): Move models to common
