+++ common/go/src/go.mod
@@ -1,3 +1,8 @@
 module powerpi/common
 
 go 1.24.1
+
+require (
+	github.com/eclipse/paho.mqtt.golang v1.5.0
+	github.com/stretchr/testify v1.10.0
+)
