+++ services/shutdown/src/go.mod
@@ -1,8 +1,6 @@
 module powerpi/shutdown
 
-go 1.23.0
-
-toolchain go1.24.1
+go 1.24.1
 
 require (
 	github.com/eclipse/paho.mqtt.golang v1.5.0
